+++ src/lib/check-uris.ts
